Deletion checks for Customer and Room (Fixes #6)
Right now, a simple error message is shown. Maybe we
could improve this by adding the option to "hide" the
entry if all reservations in which the items are referred
are hidden / in the past?

The object list would then probably need to have a way
of showing deleted entries in addition to normal ones.