RoomEditor: Enable only useful items; fix input check
The amount of people in a room can only be 1-3.
If the amount is 1, only Single makes sense.
If the amount is 2, only Single (w or w/o kid) + double makes sense.
If the amount is 3, all prices make sense.

Disable unneeded fields and only check for > 0
in enabled fields to fix sanity checking.